+++ src/App.vue
@@ -1,45 +1,19 @@
 <template>
-  <div class="signup">
-    <table>
-      <tr>
-        <th>メールアドレス：</th>
-      </tr>
-      <tr>
-        <td><input type="email" v-model="mailaddress" /></td>
-      </tr>
-      <tr>
-        <th>パスワード：</th>
-      </tr>
-      <tr>
-        <td><input type="password" v-model="password" /></td>
-      </tr>
-    </table>
-    <button @click="signUp">登録</button>
-  </div>
+  <SingUpComponent />
+  <SingInComponent />
 </template>
 
 <script>
-import { projectAuth } from "./firebase.js"
-import { createUserWithEmailAndPassword } from "firebase/auth"
+import SingUpComponent from "./components/SignUp.vue"
+import SingInComponent from "./components/SignInComponent.vue"
 
 export default {
-  data() {
-    return {
-      mailaddress: "",
-      password: "",
-    }
+  components: {
+    SingUpComponent,
+    SingInComponent,
   },
-  methods: {
-    signUp: function () {
-      createUserWithEmailAndPassword(
-        projectAuth,
-        this.mailaddress,
-        this.password
-      )
-        .then()
-        .catch()
-    },
-  },
+  data() {},
+  methods: {},
 }
 </script>
